Fixed an issue with reading data
There was an issue with how the processor reads data and how it uses that data, where it would error out and then cause the whole process to crumble.

this issue is now fixed and the processor will allways accept valid data, the processor is also capt at 255 bytes of storage per address, and the default address amount for ram is 255, since I believe currently the processor would be unable to load into them, tho it might work since it just uses the int and does not use anything else. Please also advise to change the size of the ram and rom (if needed) when using higher addresses, the preset only does upto 0xff but you can go up as high as hex will allow you, I would recommend for big projects to use 0xffff instead of 0xff.